requirements.txt
appdirs==1.4.3
boto==2.46.1
boto3==1.4.4
botocore==1.5.45
Django==1.10
django-storages==1.5.2
docutils==0.13.1
gunicorn==19.7.1
jmespath==0.9.2
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.7
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
s3transfer==0.1.10
six==1.10.0
Unidecode==0.4.20
virtualenv==15.1.0
whitenoise==3.3.0
django-smartfields==1.0.9
django-robots
python3-memcached==1.51
python-decouple==3.0
django-tinymce==2.6.0
django-mptt==0.8.7